+++ backend/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod api;
 pub mod auth;
 pub mod model;
+pub mod route;
